+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [v0.2.0] - 2024-01-22
+
+### Changed
+
+- Default export is now JavaScript eslint config.
+
+### Added
+
+- Different configuration files to JavaScript and TypeScript.
+
 ## [v0.1.3] - 2024-01-19
 
 ### Fixed